[MI-2669]: Created a hook for intercepting redux changes and moved the central login into it, formatted code
